Added view and serializer into token file
