+++ backend/pharmacy/api/views/userActions/token/serializer.py
@@ -35,3 +35,14 @@ from django.contrib.auth.models import User
 
 # Imports used in search functionality
 from rest_framework import filters
+
+class MyTokenObtainPairSerializer(TokenObtainPairSerializer):
+    @classmethod
+    def get_token(cls, user):
+        token = super().get_token(user)
+
+        token['username'] = user.username
+        token['email'] = user.email
+
+        return token
+
